refactor(parser): Rebuild the load_split_audiofile to save the chunks as mp3s
This function (previously split_audiofile) has been reworked so that it is able to save the split chunks (in any duration) as mp3 and be able to load these back in using torchaudio using pydub. Then it can save the waveform as a new MusicData instance. The detect leading silence function will need to be reworked as currently the beginning and end silences in the audio file cannot be trimmed